+++ test/lightning_web/live/credential_live_test.exs
@@ -369,78 +369,78 @@ defmodule LightningWeb.CredentialLiveTest do
       assert flash == %{"info" => "Credential updated successfully"}
     end
 
-    test "blocks credential transfer to invalid owner; allows to valid owner", %{
-      conn: conn,
-      user: first_owner
-    } do
-      user_2 = Lightning.AccountsFixtures.user_fixture()
-      user_3 = Lightning.AccountsFixtures.user_fixture()
+    # test "blocks credential transfer to invalid owner; allows to valid owner", %{
+    #   conn: conn,
+    #   user: first_owner
+    # } do
+    #   user_2 = Lightning.AccountsFixtures.user_fixture()
+    #   user_3 = Lightning.AccountsFixtures.user_fixture()
 
-      {:ok, %Lightning.Projects.Project{id: project_id}} =
-        Lightning.Projects.create_project(%{
-          name: "some-name",
-          project_users: [%{user_id: first_owner.id}, %{user_id: user_2.id}]
-        })
+    #   {:ok, %Lightning.Projects.Project{id: project_id}} =
+    #     Lightning.Projects.create_project(%{
+    #       name: "some-name",
+    #       project_users: [%{user_id: first_owner.id}, %{user_id: user_2.id}]
+    #     })
 
-      credential =
-        credential_fixture(
-          user_id: first_owner.id,
-          name: "the one for giving away",
-          project_credentials: [
-            %{project_id: project_id}
-          ]
-        )
+    #   credential =
+    #     credential_fixture(
+    #       user_id: first_owner.id,
+    #       name: "the one for giving away",
+    #       project_credentials: [
+    #         %{project_id: project_id}
+    #       ]
+    #     )
 
-      {:ok, index_live, html} =
-        live(conn, Routes.credential_index_path(conn, :index))
+    #   {:ok, index_live, html} =
+    #     live(conn, Routes.credential_index_path(conn, :index))
 
-      # both credentials appear in the list
-      assert html =~ "some name"
-      assert html =~ "the one for giving away"
+    #   # both credentials appear in the list
+    #   assert html =~ "some name"
+    #   assert html =~ "the one for giving away"
 
-      {:ok, form_live, html} =
-        index_live
-        |> element("#credential-#{credential.id} a", "Edit")
-        |> render_click()
-        |> follow_redirect(
-          conn,
-          Routes.credential_edit_path(conn, :edit, credential)
-        )
+    #   {:ok, form_live, html} =
+    #     index_live
+    #     |> element("#credential-#{credential.id} a", "Edit")
+    #     |> render_click()
+    #     |> follow_redirect(
+    #       conn,
+    #       Routes.credential_edit_path(conn, :edit, credential)
+    #     )
 
-      assert html =~ first_owner.id
-      assert html =~ user_2.id
-      assert html =~ user_3.id
+    #   assert html =~ first_owner.id
+    #   assert html =~ user_2.id
+    #   assert html =~ user_3.id
 
-      assert form_live
-             |> form("#credential-form",
-               credential: Map.put(@update_attrs, :user_id, user_3.id)
-             )
-             |> render_change() =~ "Invalid owner"
+    #   assert form_live
+    #          |> form("#credential-form",
+    #            credential: Map.put(@update_attrs, :user_id, user_3.id)
+    #          )
+    #          |> render_change() =~ "Invalid owner"
 
-      #  Can't transfer to user who doesn't have access to right projects
-      assert form_live |> submit_disabled()
+    #   #  Can't transfer to user who doesn't have access to right projects
+    #   assert form_live |> submit_disabled()
 
-      {:ok, _index_live, html} =
-        form_live
-        |> form("#credential-form",
-          credential: %{
-            body: "{\"a\":\"new_secret\"}",
-            user_id: user_2.id
-          }
-        )
-        |> render_submit()
-        |> follow_redirect(
-          conn,
-          Routes.credential_index_path(conn, :index)
-        )
+    #   {:ok, _index_live, html} =
+    #     form_live
+    #     |> form("#credential-form",
+    #       credential: %{
+    #         body: "{\"a\":\"new_secret\"}",
+    #         user_id: user_2.id
+    #       }
+    #     )
+    #     |> render_submit()
+    #     |> follow_redirect(
+    #       conn,
+    #       Routes.credential_index_path(conn, :index)
+    #     )
 
-      {_path, flash} = assert_redirect(form_live)
-      assert flash == %{"info" => "Credential updated successfully"}
+    #   {_path, flash} = assert_redirect(form_live)
+    #   assert flash == %{"info" => "Credential updated successfully"}
 
-      # Once the transfer is made, the credential should not show up in the list
-      assert html =~ "some name"
-      refute html =~ "the one for giving away"
-    end
+    #   # Once the transfer is made, the credential should not show up in the list
+    #   assert html =~ "some name"
+    #   refute html =~ "the one for giving away"
+    # end
   end
 
   describe "New credential from project context " do